Change Vec3 structure types to long double
The structure Vec3 in the linear_algebra.h file had its component types 'x', 'y', and 'z' changed from double to long double. This change provides greater precision when performing computations on the structure's data.

+++ c/include/math/linear_algebra.h
@@ -38,7 +38,7 @@ extern "C" {
  *  Member 'z' is the z component of the vector
  */
 typedef struct {
-    double x, y, z;
+    long double x, y, z;
 } Vec3;
 
 /**
